+++ src/extensions/bigbang/bigbang.go
@@ -262,7 +262,7 @@ func Skeletonize(tmpPaths *layout.ComponentPaths, c types.ZarfComponent) (types.
 		rel := filepath.Join(layout.TempDir, skelName)
 		dst := filepath.Join(tmpPaths.Base, rel)
 
-		if err := utils.CreatePathAndCopy(valuesFile, dst); err != nil {
+		if err := helpers.CreatePathAndCopy(valuesFile, dst); err != nil {
 			return c, err
 		}
 
@@ -282,7 +282,7 @@ func Skeletonize(tmpPaths *layout.ComponentPaths, c types.ZarfComponent) (types.
 		rel := filepath.Join(layout.TempDir, skelName)
 		dst := filepath.Join(tmpPaths.Base, rel)
 
-		if err := utils.CreatePathAndCopy(fluxPatchFile, dst); err != nil {
+		if err := helpers.CreatePathAndCopy(fluxPatchFile, dst); err != nil {
 			return c, err
 		}
