+++ src/internal/packager/template/yaml.go
@@ -7,7 +7,7 @@ package template
 import (
 	"regexp"
 
-	"github.com/defenseunicorns/zarf/src/pkg/utils"
+	"github.com/defenseunicorns/zarf/src/pkg/utils/helpers"
 	"github.com/defenseunicorns/zarf/src/types"
 )
 
@@ -15,7 +15,7 @@ import (
 func ProcessYamlFilesInPath(path string, component types.ZarfComponent, values Values) ([]string, error) {
 	// Only pull in yml and yaml files
 	pattern := regexp.MustCompile(`(?mi)\.ya?ml$`)
-	manifests, _ := utils.RecursiveFileList(path, pattern, false)
+	manifests, _ := helpers.RecursiveFileList(path, pattern, false)
 
 	for _, manifest := range manifests {
 		if err := values.Apply(component, manifest, false); err != nil {
