+++ src/pkg/packager/composer/pathfixer.go
@@ -7,7 +7,6 @@ package composer
 import (
 	"path/filepath"
 
-	"github.com/defenseunicorns/zarf/src/pkg/utils"
 	"github.com/defenseunicorns/zarf/src/pkg/utils/helpers"
 	"github.com/defenseunicorns/zarf/src/types"
 )
@@ -46,7 +45,7 @@ func fixPaths(child *types.ZarfComponent, relativeToHead string) {
 			composed := makePathRelativeTo(kustomization, relativeToHead)
 			// kustomizations can use non-standard urls, so we need to check if the composed path exists on the local filesystem
 			abs, _ := filepath.Abs(composed)
-			invalid := utils.InvalidPath(abs)
+			invalid := helpers.InvalidPath(abs)
 			if !invalid {
 				child.Manifests[manifestIdx].Kustomizations[kustomizeIdx] = composed
 			}
