+++ src/pkg/utils/bytes.go
@@ -12,6 +12,7 @@ import (
 	"time"
 
 	"github.com/defenseunicorns/zarf/src/pkg/message"
+	"github.com/defenseunicorns/zarf/src/pkg/utils/helpers"
 )
 
 // RoundUp rounds a float64 to the given number of decimal places.
@@ -78,7 +79,7 @@ func RenderProgressBarForLocalDirWrite(filepath string, expectedTotal int64, com
 			}
 		default:
 			// Read the directory size
-			currentBytes, dirErr := GetDirSize(filepath)
+			currentBytes, dirErr := helpers.GetDirSize(filepath)
 			if dirErr != nil {
 				message.Debugf("unable to get updated progress: %s", dirErr.Error())
 				time.Sleep(200 * time.Millisecond)
