+++ src/test/e2e/05_tarball_test.go
@@ -14,6 +14,7 @@ import (
 	"github.com/defenseunicorns/zarf/src/pkg/layout"
 	"github.com/defenseunicorns/zarf/src/pkg/message"
 	"github.com/defenseunicorns/zarf/src/pkg/utils"
+	"github.com/defenseunicorns/zarf/src/pkg/utils/helpers"
 	"github.com/defenseunicorns/zarf/src/types"
 	"github.com/stretchr/testify/require"
 )
@@ -69,7 +70,7 @@ func TestMultiPartPackage(t *testing.T) {
 	require.NoError(t, err)
 	require.Equal(t, pkgData.Bytes, fullFileInfo.Size())
 	// Ensure that the pkgData shasum was correct (should be checked during deploy as well, but this is to double check)
-	err = utils.SHAsMatch(parts[0], pkgData.Sha256Sum)
+	err = helpers.SHAsMatch(parts[0], pkgData.Sha256Sum)
 	require.NoError(t, err)
 
 	e2e.CleanFiles(parts...)
